backend/requirements.txt
alembic
apscheduler
black # local
dramatiq[redis,watch] # watch - local
Faker # local
fastapi
flake8 # local
isort # local
jinja2
loguru
natasha
psycopg2-binary
python-dotenv
pytest # local
sqlalchemy
starlette_exporter
uvicorn
